requirements.txt
aiogram
g4f
feedparser
apscheduler
sqlalchemy
psycopg2-binary
aiohttp
beautifulsoup4
Pillow
python-dotenv
markdown2